+++ src/components/HorizontalBarChart/stories/HorizontalBarChart.stories.tsx
@@ -6,43 +6,20 @@ import {
   HorizontalBarChartProps,
 } from '../HorizontalBarChart';
 
-import type {RenderTooltipContentData, Series} from '../types';
-import {getSingleColor, THEME_CONTROL_ARGS} from '../../../storybook';
-
-const LABELS = ['BCFM 2019', 'BCFM 2020', 'BCFM 2021'];
-
-function buildSeries(items: number[] | number[][]): Series[] {
-  return [
-    'Womens Leggings',
-    'Mens Bottoms',
-    'Mens Shorts',
-    'Socks',
-    'Hats',
-    'Ties',
-  ].map((name, index) => {
-    const item = items[index];
-    const array = Array.isArray(item) ? item : [item];
-    return {
-      name,
-      data: array.map((number, dataIndex) => {
-        return {rawValue: number, label: LABELS[dataIndex]};
-      }),
-    };
-  });
-}
-
-const SERIES = buildSeries([
-  [3, 4, 7],
-  [0, 0, 0],
-  [4, 5, 6],
-  [8, 15, 12],
-  [48, 8, 50],
-  [1, 5, 5],
-]);
+import type {RenderTooltipContentData} from '../types';
+import {
+  getSingleColor,
+  THEME_CONTROL_ARGS,
+  TYPE_CONTROL_ARGS,
+} from '../../../storybook';
+import {
+  buildSeries,
+  COLOR_OVERRIDE_SERIES,
+  COMPARISON_SERIES,
+  SERIES,
+} from '../../../storybook/horizontal-charts';
 
 const CONTAINER_HEIGHT = 500;
-
-const SINGLE_SERIES = buildSeries([3, 7, 4, 8, 4, 1, 4, 6]);
 
 const TOOLTIP_CONTENT = {
   empty: undefined,
@@ -80,24 +57,13 @@ export default {
     },
   },
   argTypes: {
-    series: {
+    data: {
       description:
         'A collection of named data sets to be rendered in the chart. An optional color can be provided for each series, to overwrite the theme `seriesColors` defined in `PolarisVizProvider`. Required.',
     },
     isAnimated: {
       description:
         'Whether to animate the bars when the chart is initially rendered and its data is updated. Even if `isAnimated` is set to true, animations will not be displayed for users with reduced motion preferences.',
-    },
-    isSimple: {
-      description: `Determines what options the chart is rendered with. When \`true\` the chart is:
-          <br /><br />
-            - Rendered with no xAxis lines or labels.<br />
-            - The values for each series is rendered to the right of the series.<br />
-            - No Tooltips are rendered on mouse/touch-move`,
-    },
-    isStacked: {
-      description:
-        'Changes the grouping of the bars. If `true` the bar groups will stack vertically, otherwise they will render individual bars for each data point in each group.',
     },
     renderTooltipContent: {
       description:
@@ -113,6 +79,7 @@ export default {
       },
     },
     theme: THEME_CONTROL_ARGS,
+    type: TYPE_CONTROL_ARGS,
     xAxisOptions: {
       description: 'An object used to configure the xAxis and its labels.',
       defaultValue: {
@@ -133,26 +100,16 @@ const Template: Story<HorizontalBarChartProps> = (
   );
 };
 
-const SimpleTemplate: Story<HorizontalBarChartProps> = (
-  args: HorizontalBarChartProps,
-) => {
-  return (
-    <div style={{height: CONTAINER_HEIGHT}}>
-      <HorizontalBarChart isSimple={true} {...args} />
-    </div>
-  );
-};
-
 export const Default: Story<HorizontalBarChartProps> = Template.bind({});
 
 Default.args = {
-  series: SERIES,
+  data: SERIES,
 };
 
 export const AllNegative: Story<HorizontalBarChartProps> = Template.bind({});
 
 AllNegative.args = {
-  series: buildSeries([
+  data: buildSeries([
     [-3, -4, -7],
     [-7, -1, -1],
     [-4, -5, -6],
@@ -163,31 +120,15 @@ AllNegative.args = {
   isAnimated: false,
 };
 
+export const Comparison: Story<HorizontalBarChartProps> = Template.bind({});
+
+Comparison.args = {
+  data: COMPARISON_SERIES,
+};
+
 export const ColorOverrides: Story<HorizontalBarChartProps> = Template.bind({});
 ColorOverrides.args = {
-  series: [
-    {
-      name: 'Shirt',
-      data: [
-        {rawValue: 4, color: 'red', label: 'Yesterday'},
-        {rawValue: 7, label: 'Today'},
-      ],
-    },
-    {
-      name: 'Pants',
-      data: [
-        {rawValue: 5, label: 'Yesterday'},
-        {rawValue: 6, label: 'Today'},
-      ],
-    },
-    {
-      name: 'Shoes',
-      data: [
-        {rawValue: 15, label: 'Yesterday'},
-        {rawValue: 12, label: 'Today'},
-      ],
-    },
-  ],
+  data: COLOR_OVERRIDE_SERIES,
 };
 
 export const FormattedLabels: Story<HorizontalBarChartProps> = Template.bind(
@@ -195,43 +136,16 @@ export const FormattedLabels: Story<HorizontalBarChartProps> = Template.bind(
 );
 
 FormattedLabels.args = {
-  series: SERIES,
+  data: SERIES,
   xAxisOptions: {
     labelFormatter: (value) =>
       `${value} pickled peppers and pickles and a few more things`,
   },
 };
 
-export const Simple: Story<HorizontalBarChartProps> = SimpleTemplate.bind({});
+export const Stacked: Story<HorizontalBarChartProps> = Template.bind({});
 
-Simple.args = {
-  series: SINGLE_SERIES,
-};
-
-Simple.parameters = {
-  docs: {
-    description: {
-      story: `
-The Simple component is rendered without xAxis labels, grid lines & Tooltips on mouse/touch-move.
-
-To change the color of the values displayed to the right of each bar, you can override the \`xAxis.labelColor\` value in your current theme. For more information on overriding theme values, you can refer to the <a href="https://polaris-viz.shopify.io/?path=/docs/providers-polarisvizprovider--default">PolarisVizProvider documentation</a>.
-      `,
-    },
-  },
-};
-
-export const DefaultStacked: Story<HorizontalBarChartProps> = Template.bind({});
-
-DefaultStacked.args = {
-  series: SERIES,
-  isStacked: true,
-};
-
-export const SimpleStacked: Story<HorizontalBarChartProps> = SimpleTemplate.bind(
-  {},
-);
-
-SimpleStacked.args = {
-  series: SERIES,
-  isStacked: true,
+Stacked.args = {
+  data: SERIES,
+  type: 'stacked',
 };