+++ src/components/shared/HorizontalBars/index.ts
@@ -1,0 +1,2 @@
+export {HorizontalBars} from './HorizontalBars';
+export {Label} from './components';